requirements.txt
langchain-groq
langchain-huggingface
langchain-community
youtube-transcript-api
faiss-cpu
python-dotenv
pydantic
fastapi
uvicorn
streamlit
requests
